Remove files with alarms

+++ src/main/java/com/example/motioner/AlarmController.java
@@ -2,6 +2,7 @@ package com.example.motioner;
 
 import com.example.motioner.domain.entity.Alarm;
 import com.example.motioner.infrastructure.AlarmRepository;
+import com.example.motioner.infrastructure.FileManager;
 import com.example.motioner.presentation.response.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
@@ -16,9 +17,12 @@ public class AlarmController {
 
     private final AlarmRepository repository;
 
+    private final FileManager manager;
+
     @Autowired
-    public AlarmController(AlarmRepository repository) {
+    public AlarmController(AlarmRepository repository, FileManager manager) {
         this.repository = repository;
+        this.manager = manager;
     }
 
     @GetMapping("/alarms")
@@ -40,10 +44,15 @@ public class AlarmController {
     @DeleteMapping("/alarms/{id}")
     public Response delete(@PathVariable String id) {
         try {
-            repository.deleteById(UUID.fromString(id));
-            return new Response("Successfully deleted");
-        } catch (EmptyResultDataAccessException e) {
-            return new Response("Alarm with this ID is not presented in DB");
+            Optional<Alarm> stored = repository.findById(UUID.fromString(id));
+            if (stored.isPresent()) {
+                manager.removeFile(stored.get().getFilename());
+                repository.deleteById(UUID.fromString(id));
+                return new Response("Successfully deleted");
+            }
+            return new Response("Nothing to delete");
+        } catch (Exception e) {
+            return new Response("Something went wrong");
         }
     }
 }
